+++ WHATS_NEW.md
@@ -2,6 +2,13 @@
 
 ## RosarioSIS Student Information System
 
+New in 8.1
+----------
+
+Accounting
+- Salaries: add File Attached
+
+
 New in 8.0
 ----------